+++ Backend/Entities/GraphNodes/AudioFeaturesFilters/FilterTempoNode.cs
@@ -8,7 +8,7 @@ namespace Backend.Entities.GraphNodes.AudioFeaturesFilters
 {
     public class FilterTempoNode : FilterRangeNode
     {
-        protected override double? GetValue(Track t) => t.AudioFeatures.Tempo;
+        protected override int? GetValue(Track t) => (int)t.AudioFeatures.Tempo;
         public override bool RequiresAudioFeatures => true;
     }
 }
